+++ README.md
@@ -11,18 +11,36 @@
 * 구입금액 입력
 * 지난주 당첨번호 입력
 
-##### LottoNumberFactory
+##### LottoTicketFactory
 * 랜덤으로 LottoNumber 6개 뽑는다
     * Collections.shuffle() 사용
 
 ##### LottoNumber
 * 값은 1부터 45까지 중 하나이다.
 
-##### LottoGame
-* 로또 1장(숫자 6개)당 1000원
-* 지난주 당첨번호와 발급한 로또 번호들 비교
+##### LottoTicket
+* 로또 1장 -> LottoNumber 6개
+* 당첨 번호와 비교하여 맞춘 갯수를 반환한다.
 
-##### WinningMoney
+##### LottoTickets
+* LottoTicket 여러 장을 리스트 형식으로 가지고 있는다.
+
+##### WinningLottoNumbers
+* 당첨 번호를 가지고 있는다.
+
+##### LottoMachine
+* 지급한 금액만큼 LottoTicket 여러 장 발급
+
+##### LottoGame
+* 지난주 당첨번호와 발급한 로또 번호들 모두 비교
+
+##### Money
+* 발급할 LottoTicket 갯수 계산
+    * LottoTicket 1장(숫자 6개)당 1000원
+* 수익률 계산
+    * (총 당첨 금액) / (지급한 금액) -> 소수 둘째자리까지
+
+##### MatchPrize
 * 3개 일치하면 5000원
 * 4개 일치하면 50000원
 * 5개 일치하면 1500000원
