+++ src/main/java/lotto/domain/LottoMachine.java
@@ -1,5 +1,8 @@
 package lotto.domain;
 
+import lotto.domain.numbers.LottoTicket;
+import lotto.domain.numbers.LottoTickets;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.IntStream;
