+++ src/main/java/lotto/domain/numbers/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.numbers;
 
 import java.util.Objects;
