+++ src/main/java/lotto/domain/numbers/LottoTicket.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.numbers;
+
+import lotto.domain.numbers.LottoNumber;
 
 import java.util.List;
