+++ src/main/java/lotto/domain/numbers/LottoTickets.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.numbers;
+
+import lotto.domain.numbers.LottoTicket;
 
 import java.util.List;
