+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.numbers.LottoTickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
