+++ src/test/java/lotto/domain/LottoTicketFactoryTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.numbers.LottoNumber;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
