+++ src/test/java/lotto/domain/numbers/LottoNumberTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.domain.numbers;
 
+import lotto.domain.numbers.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
